chore: refactor InlinesCollection 2
